% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonizr.R
\name{Harmonizr}
\alias{Harmonizr}
\alias{is_consistent}
\title{Harmonizr}
\usage{
is_consistent(input_field_names_vector, mandatory_fields_vector,
  case_sensitive_check = TRUE, sequence_check = FALSE)
}
\arguments{
\item{input_field_names_vector}{vector of field names extracted from dataframe, list or vector}

\item{mandatory_fields_vector}{vector of mandatory fields}

\item{case_sensitive_check}{`TRUE` to keep the case sensitive check on while matching the input fields with the mandatory fields}

\item{sequence_check}{`TRUE` to check if the order of input fileds is exactly same as mandatory fields}
}
\value{
returns TRUE if the \code{input_field_names_vector} has all \code{mandatory_fields_vector} else FALSE
}
\description{
Check whether the input field(column) names vector has all mandatory fileds or not
}
\author{
Abhinandan Satpute
}
