% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homogeni.R
\name{Homogeni}
\alias{Homogeni}
\alias{is_homogeneous}
\title{HomogenizR}
\usage{
is_homogeneous(input_field_names_vector, mandatory_fields_vector,
  case_sensitive_check = TRUE, sequence_check = FALSE)
}
\arguments{
\item{input_field_names_vector}{vector of field names extracted from dataframe, list or vector}

\item{mandatory_fields_vector}{vector of mandatory fields}

\item{case_sensitive_check}{to specify the case sensitveness criteria while matching field names}

\item{sequence_check}{to specify whether to take the order of input fileds into the consideration or not}
}
\value{
returns TRUE if the \code{input_field_names_vector} has all \code{mandatory_fields_vector} else FALSE
}
\description{
Check whether the input field(column) names vector has all mandatory fileds or not
}
\author{
Abhinandan Satpute
}
